move string conversion routines to separate module
